[Iteration-1] refact: home 쿼리 옵션 분리
